+++ app/src/main/java/com/duckduckgo/app/di/DatabaseModule.kt
@@ -32,16 +32,8 @@ class DatabaseModule {
     @Singleton
     fun provideDatabase(context: Context): AppDatabase {
         return Room.databaseBuilder(context, AppDatabase::class.java, "app.db")
-            .addMigrations(MIGRATION_1_TO_2)
+            .addMigrations(AppDatabase.MIGRATION_1_TO_2)
             .build()
     }
 
-    val MIGRATION_1_TO_2: Migration = object : Migration(1, 2) {
-        override fun migrate(database: SupportSQLiteDatabase) {
-            database.execSQL("CREATE TABLE `tabs` (`tabId` TEXT NOT NULL, `url` TEXT, `title` TEXT, PRIMARY KEY(`tabId`))")
-            database.execSQL("CREATE INDEX `index_tabs_tabId` on `tabs` (tabId)")
-            database.execSQL("CREATE TABLE `tab_selection` (`id` INTEGER NOT NULL, `tabId` TEXT, PRIMARY KEY(`id`), FOREIGN KEY(`tabId`) REFERENCES `tabs`(`tabId`) ON UPDATE NO ACTION ON DELETE SET NULL)")
-            database.execSQL("CREATE INDEX `index_tab_selection_tabId` on `tab_selection` (tabId)")
-        }
-    }
 }
