+++ app/src/main/java/com/duckduckgo/app/global/db/AppDatabase.kt
@@ -16,8 +16,10 @@
 
 package com.duckduckgo.app.global.db
 
+import android.arch.persistence.db.SupportSQLiteDatabase
 import android.arch.persistence.room.Database
 import android.arch.persistence.room.RoomDatabase
+import android.arch.persistence.room.migration.Migration
 import com.duckduckgo.app.bookmarks.db.BookmarkEntity
 import com.duckduckgo.app.bookmarks.db.BookmarksDao
 import com.duckduckgo.app.httpsupgrade.db.HttpsUpgradeDomain
@@ -49,4 +51,14 @@ abstract class AppDatabase : RoomDatabase() {
     abstract fun tabsDao(): TabsDao
     abstract fun bookmarksDao(): BookmarksDao
 
+    companion object {
+        val MIGRATION_1_TO_2: Migration = object : Migration(1, 2) {
+            override fun migrate(database: SupportSQLiteDatabase) {
+                database.execSQL("CREATE TABLE `tabs` (`tabId` TEXT NOT NULL, `url` TEXT, `title` TEXT, PRIMARY KEY(`tabId`))")
+                database.execSQL("CREATE INDEX `index_tabs_tabId` on `tabs` (tabId)")
+                database.execSQL("CREATE TABLE `tab_selection` (`id` INTEGER NOT NULL, `tabId` TEXT, PRIMARY KEY(`id`), FOREIGN KEY(`tabId`) REFERENCES `tabs`(`tabId`) ON UPDATE NO ACTION ON DELETE SET NULL)")
+                database.execSQL("CREATE INDEX `index_tab_selection_tabId` on `tab_selection` (tabId)")
+            }
+        }
+    }
 }
